Consolidate files for idy
