+++ backend/src/routes/v0/info/__init__.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+
+from .health import router as health
+
+
+router = APIRouter(tags=["Info"], prefix="/info")
+router.include_router(health)
